Integration Security

+++ gourmet/src/main/java/com/example/gourmet/Repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<appUser,Long> {
+public interface UserRepository extends JpaRepository<appUser,Integer> {
 
     appUser findByUsername(String username);
 }
